Allow no more than one default in member lists
Previously the definition of a member included an optional '*' which
allowed any member to be marked as default. In a list of members any
number of members could be marked as default. Since it only makes sense
to have one default the definition of a member has been split into two
definitions: default member and nondefault member. A member list has
been defined as a list of members involving either only nondefault
members or one default member with any number of nondefault members.